% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_obj_UMAP3}
\alias{seurat_obj_UMAP3}
\title{seurat_obj_UMAP3}
\format{
A Seurat object.
}
\usage{
data(seurat_obj_UMAP3)
}
\description{
A Seurat dataset of single cell RNA sequencing data with 3 UMAP dimesions
}
\keyword{datasets}
